Update the PUN RPCs to allow for different models depending on the choice of controller type.
Further generalization of controller type.  Note this is a placeholder for a future gesture system that isn't based on LM.  Also, we will consider using Node for our networking solution in the future.

Additionally, this moves the syncing of model state to act on an IEnumerator rather than on update.  A small optimization to allow for smoother movement of objects.

Consider what framerate is optimal for updating the objects movement.  We could update the objects on the network at 30 fps while rendering at 90 fps.  this would move the transform copies out of update and into their own coroutine.

Needs testing.