Feature/gesture abstraction (#101)
* added new gesture stuffs from joey's old branch

* reimport key changes

* removed old metas

* new editor version

* Setup scene for testing Aristo API. Aristo data is translated into body input. Body Input can create a nice 'blank' copy of itself.

* fixed leap motion input parsing.

* added a comment to the visual feedback section with old stuff noted. Couldn't quite figure out where the new
color stuff had been done. Will ask later when someone is around. Began updating the wiki with relevant
information to the grasp gesture and generally gestures.

* Issue #14 outlined an example jobsystem and componant data for the gestures

* Moved EmbodiedInput to ECS Job sytstem.  #14

* Moved EmbodiedInput to ECS Job sytstem.  #14

* Moved EmbodiedInput to ECS Job sytstem.  #14

* Moved EmbodiedInput to ECS Job sytstem.  #14

* Moved EmbodiedInput to ECS Job sytstem.  #14

* Create SystemDescription.md

* Authoring system for ECS #14

* convert vector3 to float3

* uncommited changes made tuesday

* Update BodyComponent.cs

* Update Finger.cs

* Update Hand.cs

* Update BodyInput.cs

* Naming Update

See Layers section.

* first addition of Input description

* Updates to embodiedInput system

workaround for blittable data within how bodyInput holds data. Comments, spelling and typo fixes.

* something isn't blittable........

* Began writing about the data structure part.

to be finished later. (it just got consolidated)

* Finished Describing the BodyInput data structure.

This will likely need a revisit to clean up the language used to talk about what is happening here while also making the formatting better. For now this is a start.

* fix for the contents list links

committing to test. Will commit further tests if this does not work.

* Added a quick note on Velocity

described very briefly, how velocity is determined.

* undeleting files we want which were deleted in previous commit from this pc

* this file was not committed with the previous undeleting commit. so this is a continuation of the undeleting of
files which were deleting accidentally

* end of cleanup from the undeleting of files.

* Updating Contents and testing link

to test link from contents I need to use the main version of the file and not the preview... thus the commit for testing the link to the new section.

* not sure if this is committed yet

I let this page sit open for awhile. Looks like I started writing the ECS stuff. (but didn't write much as I began commenting the files which existed as I noticed they were in need of it)

* Added comments and slight tweaks

Lots of comments have been added to the classifer files. In particular the EmbodiedUserInputCalssifier file is in need of comment review from line 235 to line 260. This section for whatever reason is particularly confusing to me.

* meta files and project settings change.

* deleted two dangling meta files, fixed some comments, commented out the character data where it is not allowed
and attempted to figure out the other two errors which are popping up for classifier data unsuccessfully. I will
definitely need to work with someone who is more familiar with these to test things with debug.logs. Someone
still needs to go over the comments I made for the lines declared in the previous commit within the eligibility
section.

* further work on gesture system

* Got closer to traditional OOP practices for classification of gestures. #14 Made unity project settings accept
vive pro.

* new method for checking finger extended

* updated Aristro to ViveHandTracking.  Updated gesture tracking #14

* uncommited changes from joey. related to debugging gestures

* removed vive hand tracking from git tracking

* Made hands no longer retain their data values when hands were not tracked. This means that gestures will stop
being activated when hands are no longer in sight. THis was achieved by setting the positions of all parts of a
hand to -10, -10, -10 and setting all fingers to be not extended. Later this could result in an issue if we
track any gesture where a fist needs to be made

* Tested things with leap motion. #14  Leap Motion had som eissues with finger extensions and it also made a lot
of gestures happen when the hands enter the view of the sensor and when they exit the view of the sensor. This
raises the question as to whether we should not activate a gesture until it has been held for 10 frames or so.
This would help with identifying user intent, make sure velocity is not negatively impacted by the previous
changes where the dummy values are used to stop gesture activation when no hand is seen, and it might reduce
issues from this leap motion issue at the start and end of a hand's tracking.

* added a line, for slider, will try to do more tonight at home

* Made the Touch Slider

The touch slider is a line renderer with a script that looks at the position of the index finger tips for each hand and when the fingertips are on the line, or close enough to be essentially on the line, it will update the position of a point on the line

* The tSlider Script is super messy right now. However the functionality is almost there. Currently I am having
issues with the local positions vs the world positions. I have some dummy red sphere at the world points which
correspond to the local points on the line. Meaning if you put your hand in between the two spheres the point
on the line will update. After this is fixed, a quick check for the bounds of the line will be needed, to stop
the sphere from going further than the line's endpoints. #14 This will be the first tool made with the new
gesture system. (I believe)

* Edited to integrated with Emboided Input system.

* set to world space

* Corrected some typos and adjusted tSlider

Scene changes to correspond with script updates. minor bug fixes #14

* prototyped a touch slider.  Only works reliably with LM, maybe driver error with Vive Pro?  #14

* added vive hand tracking to optional dependencies

* make not dependent on PHOTON

* remove Photon and other attached legacy problems.

* didn't commit with the rest for some reason

* math error

* fix lm vpro integration

* slider material

* Update dependencies.md

* commit

* Prefab for Embodied User Input System

* Added Classifier description

A few sentences to get the System description started. The wording is awkward and needs review

* Added a file for the touch slider

This includes a description of what the Tslider will do and how it was made. Elaborations on the implementation and reformatting is likely needed. Definitely something for when this gets reviewed. For now though this is a good start.

* formatting update

I forgot to use the markdown header properly

* Header type swap

I should use the preview #14

* Deleted action files as these will be implemented uniquely for each scene.

* Added Gesture definitions #14

Also updated the contents list with links to the newly added gestures

* Fixed contents links with spaces #14

Had '_' instead of '-' in the URL

* Removed statement about this being the first #14

The statement was not needed in the docs. Slightly updated the wording of the new first bullet to reflect the change in order

* Hardware section update #14

Added a statement which reiterates how hardware support is added in a general way

* Updated Vive Hand tracking term from Aristo #14

I also noted that it used to be called aristo so that commit messages and comments referring to Aristo will make sense if someone looks here

* fix code style