Created the document for GeoPlanet Actions
Geometer's Planetarium's document detailing its actions which are performable and mildly touching on the implementation. Incomplete for the moment because I need to commit this so I can reload the webpage. The input field for the file is jumping around when I type for some reason and I am hoping a reload will fix this.