All of the actions are now added #92
The descriptions definitely need to be read through as this is the first pass being completed. I have fixed some of the issues with the initial writing in this commit also. I will need someone to go through these other than me as well to be sure they are coherent.