rough draft of functions for GeoPlanet
I might also add the functionality within RSDESPin to this doc. #92 #14 
This document will hold information on how to interact with Geometer's Planetarium at a script level. This will help us in the future come back to the files and quickly see what is happening without having to go through the scripts and search for functionality. This doc will also help anyone else who wants to contribute develop new things for the environment or add new things to the environment which may use some of the old stuff or something.